+++ lib/generators/happy_seed/devise/templates/spec/mailers/previews/devise_preview.rb
@@ -1,13 +1,21 @@
 class DevisePreview < ActionMailer::Preview
   def confirmation_instructions
-    Devise::Mailer.confirmation_instructions(User.first, "tokentokentoken")
+    Devise::Mailer.confirmation_instructions(user, "tokentokentoken")
   end
 
   def reset_password_instructions
-    Devise::Mailer.reset_password_instructions(User.first, "tokentokentoken")
+    Devise::Mailer.reset_password_instructions(user, "tokentokentoken")
   end
 
   def unlock_instructions
-    Devise::Mailer.unlock_instructions(User.first, "tokentokentoken")
+    Devise::Mailer.unlock_instructions(user, "tokentokentoken")
+  end
+
+  def invitation_instructions
+    Devise::Mailer.invitation_instructions(user, "tokentokentoken")
+  end
+
+  def user
+    User.first || User.new( email: "will@happyfuncorp.com" )
   end
 end