+++ CHANGELOG.md
@@ -4,7 +4,9 @@ Version 0.0.21
   - Integrated simple_form for scaffolded code
   - Got rid of FormUser
   - Added OAuth feature specs
-  - Added premailer and HTML to devise mail templates
+  - Added premailer and HTML to html_email generator
+  - Switch to github version of devise_invitable
+  - Updated Forgotpassword and invite emails to html
   
 Version 0.0.20
   - Updated Bootstrap panels in the templates